refactor(controller): collect account converters in separate file
