+++ internal/adapter/inbound/controller/account.go
@@ -229,123 +229,6 @@ func (r *AccountReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ct
 	return r.reporter.status(ctx, natsAccount)
 }
 
-func toAPIAccountClaims(claims *nauth.AccountClaims) *v1alpha1.AccountClaims {
-	if claims == nil {
-		return nil
-	}
-
-	return &v1alpha1.AccountClaims{
-		AccountLimits:    toAPIAccountLimits(claims.AccountLimits),
-		DisplayName:      claims.DisplayName,
-		SigningKeys:      claims.SigningKeys,
-		Exports:          toAPIExports(claims.Exports),
-		Imports:          toAPIImports(claims.Imports),
-		JetStreamEnabled: claims.JetStreamEnabled,
-		JetStreamLimits:  toAPIAJetStreamLimits(claims.JetStreamLimits),
-		NatsLimits:       toAPINatsLimits(claims.NatsLimits),
-	}
-}
-
-func toAPIAccountLimits(source *nauth.AccountLimits) *v1alpha1.AccountLimits {
-	if source == nil {
-		return nil
-	}
-
-	return &v1alpha1.AccountLimits{
-		Imports:         source.Imports,
-		Exports:         source.Exports,
-		WildcardExports: source.WildcardExports,
-		Conn:            source.Conn,
-		LeafNodeConn:    source.LeafNodeConn,
-	}
-}
-
-func toAPIAJetStreamLimits(source *nauth.JetStreamLimits) *v1alpha1.JetStreamLimits {
-	if source == nil {
-		return nil
-	}
-
-	return &v1alpha1.JetStreamLimits{
-		MemoryStorage:        source.MemoryStorage,
-		DiskStorage:          source.DiskStorage,
-		Streams:              source.Streams,
-		Consumer:             source.Consumer,
-		MaxAckPending:        source.MaxAckPending,
-		MemoryMaxStreamBytes: source.MemoryMaxStreamBytes,
-		DiskMaxStreamBytes:   source.DiskMaxStreamBytes,
-		MaxBytesRequired:     source.MaxBytesRequired,
-	}
-}
-
-func toAPINatsLimits(source *nauth.NatsLimits) *v1alpha1.NatsLimits {
-	if source == nil {
-		return nil
-	}
-
-	return &v1alpha1.NatsLimits{
-		Subs:    source.Subs,
-		Data:    source.Data,
-		Payload: source.Payload,
-	}
-}
-
-func toAPIImports(imports nauth.Imports) v1alpha1.Imports {
-	result := make(v1alpha1.Imports, len(imports))
-	for i, imp := range imports {
-		result[i] = &v1alpha1.Import{
-			Account:      string(imp.AccountID),
-			Name:         imp.Name,
-			Subject:      v1alpha1.Subject(imp.Subject),
-			LocalSubject: v1alpha1.RenamingSubject(imp.LocalSubject),
-			Type:         toAPIExportType(imp.Type),
-			Share:        imp.Share,
-			AllowTrace:   imp.AllowTrace,
-		}
-	}
-	return result
-}
-
-func toAPIExports(exports nauth.Exports) v1alpha1.Exports {
-	result := make(v1alpha1.Exports, len(exports))
-	for i, exp := range exports {
-		export := v1alpha1.Export{
-			Name:                 exp.Name,
-			Subject:              v1alpha1.Subject(exp.Subject),
-			Type:                 toAPIExportType(exp.Type),
-			ResponseType:         v1alpha1.ResponseType(exp.ResponseType),
-			ResponseThreshold:    exp.ResponseThreshold,
-			AccountTokenPosition: exp.AccountTokenPosition,
-			Advertise:            exp.Advertise,
-			AllowTrace:           exp.AllowTrace,
-			Latency:              toAPIServiceLatency(exp.Latency),
-		}
-		result[i] = &export
-	}
-	return result
-}
-
-func toAPIServiceLatency(latency *nauth.ServiceLatency) *v1alpha1.ServiceLatency {
-	if latency == nil {
-		return nil
-	}
-
-	return &v1alpha1.ServiceLatency{
-		Sampling: v1alpha1.SamplingRate(latency.Sampling),
-		Results:  v1alpha1.Subject(latency.Results),
-	}
-}
-
-func toAPIExportType(exportType nauth.ExportType) v1alpha1.ExportType {
-	switch exportType {
-	case nauth.ExportTypeStream:
-		return v1alpha1.Stream
-	case nauth.ExportTypeService:
-		return v1alpha1.Service
-	default:
-		return v1alpha1.Stream
-	}
-}
-
 func (r *AccountReconciler) collectAccountResources(ctx context.Context, account *v1alpha1.Account, accountID string) (domain.AccountResources, accountAdoptionRefs, error) {
 	resources := domain.AccountResources{Account: *account}
 	refs := accountAdoptionRefs{}
@@ -523,50 +406,4 @@ func newAdoptionRef(resource metav1.ObjectMeta, observedGenerationDesiredClaim *
 		ObservedGeneration:             resource.Generation,
 		ObservedGenerationDesiredClaim: observedGenerationDesiredClaim,
 	}
-}
-
-func toAPIAdoptions(adoptions *nauth.AccountAdoptions, adoptionRefs accountAdoptionRefs) *v1alpha1.AccountAdoptions {
-	if adoptions == nil {
-		return nil
-	}
-
-	result := &v1alpha1.AccountAdoptions{}
-	for _, adpRef := range adoptionRefs.exports {
-		var status v1alpha1.AccountAdoptionStatus
-		adpResult := adoptions.Exports.Get(adpRef.Ref)
-		if adpResult != nil && adpResult.IsSuccessful() {
-			status = v1alpha1.AccountAdoptionStatus{
-				Status:                         metav1.ConditionTrue,
-				Reason:                         conditionReasonOK,
-				Message:                        "Adopted",
-				DesiredClaimObservedGeneration: adpRef.ObservedGenerationDesiredClaim,
-			}
-		} else {
-			status = v1alpha1.AccountAdoptionStatus{
-				Status:                         metav1.ConditionFalse,
-				Reason:                         conditionReasonNOK,
-				DesiredClaimObservedGeneration: adpRef.ObservedGenerationDesiredClaim,
-			}
-			if adpResult == nil {
-				if adpRef.ObservedGenerationDesiredClaim == nil {
-					status.Message = "Adoption pending: no desired claim"
-				} else {
-					status.Message = "WARN: No adoption result reported"
-				}
-			} else if failure := adpResult.Failure; failure != "" {
-				status.Reason = string(failure)
-				status.Message = adpResult.Message
-			} else {
-				status.Message = "Adopted"
-			}
-		}
-		result.Exports = append(result.Exports, v1alpha1.AccountAdoption{
-			Name:               adpRef.Name,
-			UID:                adpRef.UID,
-			ObservedGeneration: adpRef.ObservedGeneration,
-			Status:             status,
-		})
-	}
-
-	return result
 }
